+++ NEWS.md
@@ -1,4 +1,4 @@
-# ospsuite.utils 1.1.0.9000
+# ospsuite.utils 1.2.0
 
 MAJOR CHANGES